Add files via upload
For those who are still using OS 10.15 on the Mac and are therefore excluded from qt6. The last version chosen before pyqt6 was used plus various “cherry picks” bug fixes that seemed important to me. In addition, a few small changes such as selecting the "Fusion" style and a small change to the windows with "topmost" to prevent them from sliding backwards.

pip3 install .

is working

./build-macos-app.sh

is also working, but the resulting file is too big to upload.


Maybe one or the other has a need...